% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdasap_app.R
\name{bdasap_app}
\alias{bdasap_app}
\title{Run the bdasap application}
\usage{
bdasap_app(runApp = TRUE)
}
\arguments{
\item{runApp}{should the app be run on completion, defaults to TRUE}
}
\description{
Run the bdasap application
}
