% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-controls.R
\name{controlsServer}
\alias{controlsServer}
\title{control server}
\usage{
controlsServer(id, data)
}
\arguments{
\item{id}{unique identifier to match the controls UI}

\item{data}{data to use for populating the choices
and selections of the pickers}
}
\description{
control server
}
