% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-controls.R
\name{controlsUI}
\alias{controlsUI}
\title{Controls UI module}
\usage{
controlsUI(id)
}
\arguments{
\item{id}{unique identifier of the module}
}
\description{
two multiselect controls
one for treatment and the other for parameters
}
