% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot.R
\name{make_plot}
\alias{make_plot}
\title{Make Lab Plot}
\usage{
make_plot(data, trta, param)
}
\arguments{
\item{data}{data.frame to use for the plot}

\item{trta}{treatments to average for each line over time}

\item{param}{parameters to facet the plot by}
}
\value{
a plotly object to be used inside the app
}
\description{
A faceted parameter plot 
with an averaged line for each treatment 
across the course of a study
}
