% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-make_table.R
\name{make_table}
\alias{make_table}
\title{Make Lab Table}
\usage{
make_table(data, trta, param)
}
\arguments{
\item{data}{data.frame to use for the table}

\item{trta}{treatments to include in the table as columns}

\item{param}{parameters to include in the table as rows}
}
\value{
a data.frame summary to be used in the app
}
\description{
A table of laboratory values
}
