% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-plot.R
\name{plotServer}
\alias{plotServer}
\title{Plot module server}
\usage{
plotServer(id, data)

plotServer(id, data)
}
\arguments{
\item{id}{matching unique namespace ID of the UI module}

\item{data}{data used for plotting and controls}
}
\description{
controls subset the plot and change the subtitle of the module

controls subset the plot and change the subtitle of the module
}
