% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-read_data_dbplyr.R
\name{read_data_dbplyr}
\alias{read_data_dbplyr}
\title{Read Data}
\usage{
read_data_dbplyr()
}
\value{
a list of data.frames for use in the application
}
\description{
Function to read in data needed for application
}
\examples{
read_data()
}
