% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-table.R
\name{tableServer}
\alias{tableServer}
\title{Table Server code}
\usage{
tableServer(id, data)
}
\arguments{
\item{id}{matching namespace of UI function}

\item{data}{data to display in the table, passed to make_table}
}
\description{
Table Server code
}
