% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-write_inputs.R
\name{write_inputs}
\alias{write_inputs}
\title{Write Inputs}
\usage{
write_inputs(lst)
}
\arguments{
\item{lst}{list of reactive values from the application}
}
\description{
Function to write the user session info
}
\examples{
write_inputs(all_inputs)
}
