% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Make Lab Table}
\usage{
make_table(data, trta, paramcd)
}
\arguments{
\item{data}{data.frame to use for the table}

\item{trta}{treatments to include in the table as columns}

\item{param}{parameters to include in the table as rows}
}
\value{
a ggplot2 object to be used inside the app
}
\description{
A table of laboratory values
}
